refactor(swarm-node): move origin push onto the dispatch engine behind discipline-carrying bounds
The engine owned retrieval dispatch while origin push hand-rolled its own
loop in the provider, reaching through pub(crate) engine accessors for the
topology, client handle and ordering. Rename RetrievalEngine to
DispatchEngine, move the sequential push walk (candidate banding, receipt
depth verdicts, shallow-receipt scoring, the verdict precedence lattice)
into the engine, and delete the accessors the provider no longer needs.

RaceBounds gains the racing and sequential constructors that carry the
discipline-to-commit-point rule: racing dispatch requires
commit-at-dispatch (a losing raced leg is cancelled by drop and a dropped
reservation releases), only sequential dispatch may pair with an on-verify
commit. The bin-route primary becomes RaceBounds::sequential; the fallback
phases become RaceBounds::racing. Behaviour is unchanged throughout.
